+++ sdk/java/src/test/java/io/github/piyushmishra/webweavex/TestJavaSDK.java
@@ -1,3 +1,5 @@
+package io.github.piyushmishra.webweavex;
+
 public class TestJavaSDK {
   public static void main(String[] args) {
     WebWeaveXClient client = new WebWeaveXClient("http://127.0.0.1:8001");